libc: Reduce the size of `id_t`.
This doesn't need to be any bigger than `pid_t`, `gid_t`, and `uid_t`, since it
doesn't need to store a discriminant along with the ID.

+++ libc/include/sys/types.h
@@ -54,7 +54,7 @@ typedef  suseconds_t; */
 typedef uint64_t time_t;
 typedef uint16_t timer_t;
 
-typedef uint64_t id_t;
+typedef uint32_t id_t;
 typedef uint32_t gid_t;
 typedef uint32_t uid_t;
 typedef uint32_t pid_t;
